% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectramix_function.R
\name{smix}
\alias{smix}
\title{create a new weighted spectrum}
\usage{
smix(data, weights, normalize = FALSE)
}
\arguments{
\item{data}{a data frame containing the spectra to mix. it is assumed the first column
contains the Wavelength labels}

\item{weights}{a 2-col matrix of the spectra to mix. The first column is the column index of
the spectrum in data, the second column is the weight of the spectrum}

\item{normalize}{if the weights should be normalized}
}
\value{
a data frame containing with the wavelength labels in the first column and the mixed spectrum in the second
}
\description{
create a new spectrum from numeric weights. the weights can add up to 1 or 100 indicating
a plain factor or percentages. if they add up to some other value, indicate via normalize = TRUE,
that you want a normalization.
}
\examples{
 \donttest{ }
}
