% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectramix_function.R
\name{smix_by_names}
\alias{smix_by_names}
\title{create a new weighted spectrum with spectrum names}
\usage{
smix_by_names(data, weights_df, normalize = FALSE)
}
\arguments{
\item{data}{a data frame containing the spectra to mix. it is assumed the first column
contains the Wavelength labels}

\item{weights_df}{a 1-row data frame with the weights. the column names are the names of the spectra
to be mixed from data. the row itself contains its weights}

\item{normalize}{if the weights should be normalized}
}
\value{
a data frame containing with the wavelength labels in the first column and the mixed spectrum in the second
}
\description{
a convenience function to create a new spectrum from named weights. the weights can add up to 1 or 100 indicating
a plain factor or percentages. if they add up to some other value, indicate via normalize = TRUE,
that you want a normalization.
}
