% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectramix_function.R
\name{smix_filter}
\alias{smix_filter}
\title{filter a spectrum by a filter config}
\usage{
smix_filter(spectrum, filter)
}
\arguments{
\item{spectrum}{the spectrum to filter}

\item{filter}{the filter. right now it is just a vector containing the bandwidth
labels from spectrum}
}
\value{
a new spectrum with just the labels and values after filtering
}
\description{
filter a spectrum by a filter config
}
\examples{
 \donttest{ }
}
